updated missions  pages and reducer
