--- src/pages/Profile.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Profile = () => (
-  <div className="myProfile-container" />
-);
-
-export default Profile;
