--- src/redux/configureStore.js
@@ -1,13 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import logger from 'redux-logger';
-import missions from './missionsReducer';
-
-const store = configureStore({
-  reducer: {
-    missions,
-  },
-
-  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
-});
-
-export default store;
